Generate tar.gz archive file with results & web artifacts.
After running all the test cases, a compressed file will be created with
all the results files and web artifacts to review them. The file has a
UTC date-time prefix and looks like this:

20230620-110654-cnf-test-results.tar.gz

That "20230620-110654" june-20th 2023, 11:06:54.

This is an extra file that will be created in the output folder, so it
won't remove any of the files that are zipped from the output folder. It
will be created by default at the end, but its generation can be
skipped/omitted setting the env var TNF_OMIT_ARTIFACTS_FILE to "true".

This is the content of the results.tar.gz file:
- claim.json
- cnf-certification-tests_junit.xml
- claimjson.js
- classification.js
- results-embed.html
- results.html

The files classification.js, results-embed.html and results.html have
been moved into the cnf-certification-test/html (new) folder so they can
be used by the golang's embed package.

As the claimjson.js is just the same content as the claim.json file
prefixed by a "var initjson = " text, I removed the bash script that
creates it and made it in go.

ToDos in further PRs:
- The classification.js file can (easily) also be built dynamically in
  go after the test suite has finished and added to the gz file.
- Include the execution log (tnf-execution.log) in the tar.gz file.
- UTs